move all functions into archive_retrieve.py per feedback
