+++ archive_retrieve.py
@@ -1,5 +1,7 @@
 import paho.mqtt.subscribe as subscribe
-import json, ssl, sqlite3, av, hashlib, os
+import json, ssl, sqlite3, av, hashlib, os, requests, config, discord
+from bs4 import BeautifulSoup
+from printer_config import PRINTERS
 
 db_file = "printers.sqlite"
 printer_table = "print_status"
@@ -164,3 +166,61 @@ def get_status_from_mqtt(printer, printer_id):
     return status
 
 
+# thanks https://plainenglish.io/blog/send-an-embed-with-a-discord-bot-in-python
+async def send_printer_status(message):
+    database = get_database_handle()
+
+    for printer_id in PRINTERS:
+        status = get_status_from_db(printer_id, database)
+        if status is not False:
+            printer = status["printer"]
+            embed = discord.Embed(
+                title="🖨 " + printer + ": " + status["state"].title(),
+                color=0xFF5733
+            )
+
+            value = ("""`{0}`\n{1} Min Remain""".format(status["job"], status["mins"]))
+            image_path  = "/tmp/" + status["job_hash"] + ".jpg" # todo - avoid writing to disk
+            with open(image_path, 'wb') as file:
+                file.write(status["image"])
+            file = discord.File(image_path, filename="printer.jpg")
+            embed.set_thumbnail(url="attachment://printer.jpg")
+        else:
+            value = "Failed to get status for printer " + printer_id
+            file = None
+            image_path = None
+            printer = None
+
+        embed.add_field(name=printer, value=value, inline=False)
+        await message.channel.send(embed=embed, file=file)
+        if image_path is not None:
+            os.remove(image_path)
+
+    database.close()
+
+
+def get_shop_hours():
+    r = requests.get(config.SHOP_HOURS_URL)
+    soup = BeautifulSoup(r.text, 'html.parser')
+    shop_hours = soup.find(id="shophours")
+
+    hours_dict = {}
+
+    for tr in shop_hours.find_all('tr'):
+        td = tr.find_all("td")
+        hours_dict[td[0].text] = td[1].text
+
+    markdown_string = f'```\nCurrent Shop Hours (fetched from https://synshop.org/hours) \n===\n'
+
+    for k,v in hours_dict.items():
+        spaces = ""
+        s = 11 - (len(k) + 1)
+        for x in range(s):
+            spaces = spaces + " "
+
+        markdown_string += f'{k}:{spaces}{v}\n'
+
+    markdown_string += f'\n{config.SHOP_ADDRESS}\n'
+    markdown_string += f'\n{config.MEMBERSHIP_NOTICE}\n```'
+
+    return markdown_string
